move some personality data from asm to C
