+++ data/data_80F4278.s
@@ -1,32 +1,5 @@
         .section .rodata
 
-	.global gUnknown_80F4278
-gUnknown_80F4278: @ 80F4278
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x03, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-
-	.global gUnknown_80F4290
-gUnknown_80F4290: @ 80F4290
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x06, 0x00, 0x00, 0x00
-        .byte 0x02, 0x00, 0x02, 0x00
-        .byte 0x09, 0x00, 0x0B, 0x00
-        .byte 0x0D, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-
-	.global gUnknown_80F42A8
-gUnknown_80F42A8: @ 80F42A8
-        .byte 0x00, 0x00, 0x00, 0x00
-        .byte 0x05, 0x00, 0x00, 0x00
-        .byte 0x0E, 0x00, 0x04, 0x00
-        .byte 0x05, 0x00, 0x05, 0x00
-        .byte 0x05, 0x00, 0x00, 0x00
-        .byte 0x00, 0x00, 0x00, 0x00
-
 	.global gUnknown_80F42C0
 gUnknown_80F42C0: @ 80F42C0
 	.string "Pokémon\0"
