+++ ld_script.txt
@@ -110,9 +110,9 @@ SECTIONS {
         src/pokemon_mail.o(.text);
         src/code_803C1B4.o(.text);
         asm/code_803C1D0.o(.text);
-        src/personality_test_pre.o(.text);
-        asm/personality_test_2.o(.text);
         src/personality_test.o(.text);
+        asm/personality_test_2.o(.text);
+        src/personality_test_1.o(.text);
         asm/code_803D110.o(.text);
         src/code_8040094.o(.text);
         asm/code_8040094.o(.text);
@@ -228,8 +228,8 @@ SECTIONS {
         data/data_80B9BB8.o(.rodata);
         data/pokemon_mail.o(.rodata);
         data/personality_test.o(.rodata);
-        src/personality_test_pre.o(.rodata);
         src/personality_test.o(.rodata);
+        src/personality_test_1.o(.rodata);
         data/data_80F4278.o(.rodata);
         src/exclusive_pokemon.o(.rodata);
         data/data_810AE24.o(.rodata);
