+++ src/data/nature_description.h
@@ -1,0 +1,232 @@
+ALIGNED(4) const char gQuirkyDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The quirky type.#P"
+    "#+People consider you an eccentric\n"
+    "#+who does things at your own pace.#P"
+    "#+You never break tempo.#P"
+    "#+Your carefree nature makes you\n"
+    "#+attractive. But you also happen\n"
+    "#+to be somewhat childish...#P"
+    "#+You are fickle and cause problems for the\n"
+    "#+people who have to go along with you.#P"
+    "#+Maybe people are even upset\n"
+    "#+with you for being so fickle?#P"
+    "#+If you realize how selfishly\n"
+    "#+you~27re behaving~2c try to think\n"
+    "#+before you do anything rash.#P"
+    "#+A quirky person like you should be...";
+
+ALIGNED(4) const char gLonelyDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The lonely type.#P"
+    "#+You always act cheerful and\n"
+    "#+jocular around other people.#P"
+    "#+But that~27s only because\n"
+    "#+you are with other people.#P"
+    "#+However~2c when you get alone...#P"
+    "#+Do you find yourself\n"
+    "#+feeling oddly depressed?#P"
+    "#+That~27s why you always\n"
+    "#+want to be with others.#P"
+    "#+But if you go around feeling\n"
+    "#+depressed too much...#P"
+    "#+Your nutritional balance goes out\n"
+    "#+of wack. Eat more vegetables!#P"
+    "#+However...#W\n"
+    "#+It~27s not a bad thing\n"
+    "#+to feel lonely.#P"
+    "#+You know what it~27s like when you aren~27t\n"
+    "#+alone~2c so that~27s why you feel lonesome.#P"
+    "#+And that~27s why you aren~27t really alone.#P"
+    "#+A lonely person like you should be...";
+
+ALIGNED(4) const char gRelaxedDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The relaxed type.#P"
+    "#+Do you occasionally\n"
+    "#+zone out and miss a bus?#P"
+    "#+Or do you find yourself dozing off?#P"
+    "#+Or is your reaction time\n"
+    "#+a little slower than others?#P"
+    "#+But that~27s not necessarily\n"
+    "#+a bad thing.#P"
+    "#+You can do things at your own\n"
+    "#+tempo without feeling pressured.#P"
+    "#+You can live in a relaxed and\n"
+    "#+unhurried manner without worries.#P"
+    "#+I think that~27s a happy\n"
+    "#+lifestyle to be envied~2c even.#P"
+    "#+You~27re also surprisingly popular.#P"
+    "#+The way you vacantly stare\n"
+    "#+off into the distance...#P"
+    "#+It should make that someone\n"
+    "#+special~27s pulse race.#P"
+    "#+A relaxed person like you should be...";
+
+ALIGNED(4) const char gCalmDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The calm type.#P"
+    "#+You~27re capable of giving advice\n"
+    "#+to friends with worries.#P"
+    "#+You don~27t like to fight.#P"
+    "#+You~27re a warm~2c kindhearted\n"
+    "#+person who cares.#P"
+    "#+You must have many friends\n"
+    "#+who look up to you.#P"
+    "#+However...#W\n"
+    "#+You may also be somewhat gullible...#P"
+    "#+As well as a little careless...#P"
+    "#+And even a little sloppy.#P"
+    "#+You might want to keep\n"
+    "#+those points in mind.#P"
+    "#+A calm person like you should be...";
+
+ALIGNED(4) const char gSassyDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The sassy type.#P"
+    "#+You tend to be somewhat cynical.#P"
+    "#+Despite that~2c there is something\n"
+    "#+appealing and lovable about you.#P"
+    "#+But do you occasionally say something\n"
+    "#+arrogant that angers others?#P"
+    "#+Do you make that mistake?#P"
+    "#+Or have people called you\n"
+    "#+conceited~2c vain~2c or selfish?#P"
+    "#+Have people said that about you?#P"
+    "#+Huh?#W You~27re telling me to get lost?#P"
+    "#+Why~2c you... Come here and say that!#W\n"
+    "#+......Gasp!#P"
+    "#+...I~27m sorry.\n"
+    "#+I let my feelings run away.\n"
+    "#+I truly regret this~2c really.#P"
+    "#+Anyway~2c your cool and aloof\n"
+    "#+attitude is what defines you.#P"
+    "#+It makes you exasperating and\n"
+    "#+appealing at the same time.#P"
+    "#+A sassy person like you should be...";
+
+ALIGNED(4) const char gHastyDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The hasty type.#P"
+    "#+You like to take charge\n"
+    "#+and get things done.#P"
+    "#+You~27re a real go-getter.#P"
+    "#+But are you also stressed out?#P"
+    "#+You get irritated when your\n"
+    "#+friends don~27t show up on time.#P"
+    "#+You get frustrated when things\n"
+    "#+don~27t turn out the way you expect.#P"
+    "#+Maybe you jab the elevator button\n"
+    "#+if the elevator is slow to arrive.#P"
+    "#+...Maybe you~27re already jabbing\n"
+    "#+the A Button repeatedly now.#P"
+    "#+Beware--getting too easily irritated\n"
+    "#+just isn~27t good for your well-being.#P"
+    "#+A hasty person like you should be...";
+
+ALIGNED(4) const char gTimidDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The timid type.#P"
+    "#+You may find it hard to go\n"
+    "#+to the washroom at night.#P"
+    "#+You may also find it too frightening\n"
+    "#+to go back to school to get something\n"
+    "#+you left behind in class.#P"
+    "#+If you~27re ever walking on a dark\n"
+    "#+street at night~2c you probably turn\n"
+    "#+around often to check behind you.#P"
+    "#+But your timid nature is\n"
+    "#+also your good point!#P"
+    "#+Because those who know fear are\n"
+    "#+those who know true courage.#P"
+    "#+A timid person like you should be...";
+
+ALIGNED(4) const char gNaiveDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The naive type.#P"
+    "#+You are highly curious~2c\n"
+    "#+and you love rare things.#P"
+    "#+Your cheerful and carefree\n"
+    "#+attitude should make things fun\n"
+    "#+for the people around you.#P"
+    "#+But you do have one flaw.\n"
+    "#+You can be childish.#P"
+    "#+You can never sit still.\n"
+    "#+You~27re always on the move.#P"
+    "#+You can also be selfish~2c\n"
+    "#+so you should watch yourself.#P"
+    "#+A naive person like you should be...";
+
+ALIGNED(4) const char gImpishDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The impish type.#P"
+    "#+You~27re playful~2c cheerful~2c\n"
+    "#+and you love pranks.#P"
+    "#+You~27re also kindhearted.#P"
+    "#+That~27s why the people around\n"
+    "#+you find you so irresistible.#P"
+    "#+You must be the most\n"
+    "#+popular person around!#P"
+    "#+Oh?#W You~27re not that popular?#W\n"
+    "#+You~27re either being modest...\n"
+    "#+or you just don~27t notice it.#P"
+    "#+I bet people are just too shy\n"
+    "#+to let their feelings be known.#P"
+    "#+There~27s someone out there who~27s\n"
+    "#+afraid to declare their love for you!#P"
+    "#+An impish person like you should be...";
+
+ALIGNED(4) const char gJollyDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The jolly type.#P"
+    "#+Always laughing and smiling~2c\n"
+    "#+you uplift everyone around you.#P"
+    "#+You love jokes!#P"
+    "#+You have lots of friends~2c and\n"
+    "#+you~27re popular wherever you go.#P"
+    "#+But sometimes you get carried away\n"
+    "#+and say things that get you in trouble.#P"
+    "#+You should learn to think before\n"
+    "#+saying or doing anything.#P"
+    "#+A jolly person like you should be...";
+
+ALIGNED(4) const char gBraveDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The brave type.#P"
+    "#+You have a strong sense of justice.#W\n"
+    "#+You hate evil.#W\n"
+    "#+You will take on any opponent.#P"
+    "#+You are truly a hero!#P"
+    "#+Go forth!#P"
+    "#+For justice...#W\n"
+    "#+For peace on earth...#W\n"
+    "#+Fight the forces of evil!#P"
+    "#+......#WIf I~27m wrong...#W\n"
+    "#+Work at becoming a true hero!#P"
+    "#+A brave person like you should be...";
+
+ALIGNED(4) const char gDocileDescription[] = 
+    "#+You appear to be...#W\n"
+    "#+The docile type.#P"
+    "#+You~27re very kindhearted.#W\n"
+    "#+Very helpful.#W\n"
+    "#+You can make friends with anyone.#P"
+    "#+You~27re a wonderful person.#P"
+    "#+......#WIs that going overboard?#W\n"
+    "#+I don~27t think so.#P"
+    "#+You yourself should be\n"
+    "#+the best judge of that.#P"
+    "#+A docile person like you should be...";
+
+
+ALIGNED(4) const char gHardyDescription[] =
+    "#+You appear to be...#W\n"
+    "#+The hardy type.#P#+You do your homework diligently~2c\n"
+    "#+and you know to eat properly.#P"
+    "#+You have strong willpower that lets\n"
+    "#+you complete tasks~2c however tough.#P"
+    "#+But~2c you can also be stubborn to the\n"
+    "#+point of even feuding with friends...#P"
+    "#+Nothing will go right for you when you~27re\n"
+    "#+irritated~2c so learn to laugh it off.#P"
+    "#+A hardy person like you should be...";
